+++ src/cpp/tests/test_tardigrade_hydraMicromorphicDruckerPragerPlasticity.cpp
@@ -117,6 +117,204 @@ bool tolerantCheck( const double &v1, const double &v2, double eps = 1e-6, doubl
 
 }
 
+BOOST_AUTO_TEST_CASE( test_maxDeltaPlasticDeformation, * boost::unit_test::tolerance( DEFAULT_TEST_TOLERANCE ) ){
+
+    // Form a hydra class and residual class
+    floatType time = 1.23;
+
+    floatType deltaTime = 2.34;
+
+    floatType temperature = 3.45;
+
+    floatType previousTemperature = 4.56;
+
+    variableVector deformationGradient = {  1.00157757,  0.00159138,  0.00672005,
+                                            0.01747159,  1.01122277,  0.00555118,
+                                            0.01112217, -0.00885205,  0.99308943 };
+
+    variableVector microDeformation = { 0.99460588, -0.0078411 ,  0.01145249,
+                                       -0.00307139,  0.97798389, -0.00509779,
+                                        0.01189977, -0.01587541,  0.98377259 };
+
+    variableVector gradientMicroDeformation = { -1.35868385e-02, -1.03142977e-02,  6.54880619e-03, -2.03947530e-02,
+                                                -3.31494137e-03, -3.45686183e-03, -3.15745117e-04, -3.70848549e-03,
+                                                -9.38693885e-03, -3.68243465e-03,  1.96694582e-02,  2.22080009e-02,
+                                                 9.18337942e-05,  6.19764759e-03, -1.92190802e-02, -9.13572591e-03,
+                                                -4.25868940e-03,  1.83154579e-02, -1.24772317e-02, -8.48286787e-04,
+                                                 2.42779893e-02,  9.74255963e-04,  5.64472629e-03, -1.89685667e-02,
+                                                 1.63170400e-02,  5.15300642e-03,  2.25340032e-03 };
+
+    floatVector previousDeformationGradient = { 9.94656270e-01,  4.82152400e-02,  3.31984800e-02,  2.81918700e-02,
+         1.02086536e+00, -1.77592100e-02, -2.24798000e-03, -1.28410000e-04,
+         9.77165250e-01 };
+
+    floatVector previousMicroDeformation = { 0.96917405, -0.01777599,  0.00870406, -0.02163002,  0.9998683 ,
+        -0.01669352,  0.03355217,  0.04427456,  1.01778466 };
+
+    floatVector previousGradientMicroDeformation = { 0.05043761,  0.02160516, -0.0565408 ,  0.01218304, -0.05851034,
+         0.00485749, -0.00962607, -0.03455912,  0.04490067,  0.01552915,
+        -0.02878364,  0.00595866,  0.04750406, -0.02377005, -0.05041534,
+        -0.02922214,  0.06280788,  0.02850865, -0.00226005,  0.0146049 ,
+         0.01560184,  0.03224767,  0.05822091, -0.05294424, -0.03518206,
+         0.01831308,  0.03774438 };
+
+    floatVector previousStateVariables = {-0.02495446, -0.00169657,  0.04855598,  0.00194851,  0.01128945,
+       -0.03793713,  0.03263408,  0.01030601,  0.0045068 , -0.01572362,
+       -0.01958792, -0.00829778,  0.01813008,  0.03754568,  0.00104223,
+        0.01693138,  0.00859366,  0.01249035,  0.01746891,  0.03423424,
+       -0.0416805 ,  0.02636828, -0.02563336, -0.0305777 ,  0.0072457 ,
+       -0.04042875,  0.03853268,  0.0127249 ,  0.02234164, -0.04838708,
+        0.00944319,  0.00567852, -0.03410404, -0.03469295,  0.01955295,
+       -0.01812336,  0.01919703,  0.00543832, -0.01110494,  0.04251325,
+        0.034167  , -0.01426024, -0.04564085, -0.01952319, -0.01018143,
+        0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0 };
+
+    floatVector parameters = { 2, 0.53895133, 0.37172145,
+                               2, 0.37773052, 0.92739145,
+                               2, 0.53186824, 0.75454313,
+                               3, 0.95338442, 0.74042148, 0.09916127,
+                               3, 0.38093104, 0.49241325, 0.46187452,
+                               5, 0.82121039, 0.90566759, 0.50466975, 0.04830311, 0.85951495,
+                               3, 0.01166325, 0.05331896, 0.28081774,
+                               3, 0.32982199, 0.60161431, 0.33157768,
+                               5, 0.58881096, 0.11473813, 0.58001078, 0.83382529, 0.3260217 };
+
+    floatVector unknownVector = {  4.31202550e-01, -1.78960429e-01, -6.17986089e-01,  9.34988614e-01,
+                                   3.01500733e-01,  7.30919703e-01, -9.49515284e-01, -4.66188370e-01,
+                                   4.14220065e-03, -8.65102730e-01,  9.86066522e-01, -5.27075208e-01,
+                                  -2.51415635e-01, -5.71976170e-01, -7.89108268e-01, -5.35040429e-01,
+                                  -3.98779729e-01,  2.68884536e-01, -4.37530437e-01, -2.75446478e-01,
+                                  -9.88114313e-01, -2.68561748e-01,  6.77719634e-02, -6.75968326e-01,
+                                   1.94866217e-01, -4.13695063e-01,  2.64100990e-01, -9.47606789e-01,
+                                   7.75186921e-01, -9.67762739e-01, -7.46083938e-01,  5.54324923e-01,
+                                  -9.08209536e-01,  4.21997387e-01,  9.42092281e-01,  7.43365866e-01,
+                                   4.20323303e-01,  9.17019486e-01, -1.40373324e-01,  7.45757829e-01,
+                                  -2.88084664e-01,  8.59527306e-01, -7.02444688e-01,  8.80058030e-01,
+                                   6.65432395e-01,  1.01730274e+00, -1.88038495e-02,  4.82434492e-03,
+                                  -2.41803760e-02,  1.01105922e+00, -2.46131243e-02, -2.07588861e-02,
+                                  -1.37250795e-02,  1.01875623e+00,  9.93178816e-01,  1.99799676e-03,
+                                   3.40516069e-03, -1.37268320e-02,  1.00360734e+00,  8.04758975e-03,
+                                  -1.00877303e-02, -4.06865705e-03,  9.97654446e-01,  2.16175331e-02,
+                                   4.37468737e-03,  2.24126186e-02,  2.80173769e-03,  2.80710425e-05,
+                                  -2.48233895e-02, -9.55547806e-04,  2.13727499e-02, -1.50817155e-02,
+                                  -2.23954433e-02, -4.66105533e-03, -6.38017597e-03,  1.78576529e-02,
+                                  -2.36694442e-02,  2.10074615e-02,  9.04514995e-03,  2.02112997e-02,
+                                   5.37645354e-03,  1.55976656e-02, -8.22280632e-03, -7.52168860e-03,
+                                  -5.50628848e-03,  1.27398541e-02, -6.53544128e-03, -1.28890097e-02,
+                                   2.18834178e-02,  2.04005542e-02,
+                                   0.01, 0.02, 0.03, 0.04, 0.05,
+                                   0.06, 0.07, 0.08, 0.09, 0.10 };
+
+    unsigned int numConfigurations = 2;
+
+    unsigned int numNonLinearSolveStateVariables = 10;
+
+    std::vector< unsigned int > stateVariableIndices = { 0, 1, 2, 3, 4, 5 };
+
+    unsigned int dimension = 3;
+
+    unsigned int configuration_unknown_count = 45;
+
+    floatType tolr = 1e-2;
+
+    floatType tola = 1e-3;
+
+    unsigned int maxIterations = 24;
+
+    unsigned int maxLSIterations = 45;
+
+    floatType lsAlpha = 2.3;
+
+    class stressMock : public tardigradeHydra::residualBaseMicromorphic{
+
+        public:
+
+            using tardigradeHydra::residualBaseMicromorphic::residualBaseMicromorphic;
+
+    };
+
+    class residualMock : public tardigradeHydra::micromorphicDruckerPragerPlasticity::residual{
+
+        public:
+
+            using tardigradeHydra::micromorphicDruckerPragerPlasticity::residual::residual;
+
+    };
+
+    class hydraBaseMicromorphicMock : public tardigradeHydra::hydraBaseMicromorphic{
+
+        public:
+
+            using tardigradeHydra::hydraBaseMicromorphic::hydraBaseMicromorphic;
+
+            stressMock elasticity;
+
+            residualMock plasticity;
+
+            std::vector< unsigned int > stateVariableIndices = { 0, 1, 2, 3, 4 };
+
+            floatVector plasticParameters = { 2, 0.53895133, 0.37172145,
+                                              2, 0.37773052, 0.92739145,
+                                              2, 0.53186824, 0.75454313,
+                                              3, 0.95338442, 0.74042148, 0.09916127,
+                                              3, 0.38093104, 0.49241325, 0.46187452,
+                                              5, 0.82121039, 0.90566759, 0.50466975, 0.04830311, 0.85951495,
+                                              3, 0.01166325, 0.05331896, 0.28081774,
+                                              3, 0.32982199, 0.60161431, 0.33157768,
+                                              5, 0.58881096, 0.11473813, 0.58001078, 0.83382529, 0.3260217 };
+
+        private:
+
+            using tardigradeHydra::hydraBaseMicromorphic::setResidualClasses;
+
+            virtual void setResidualClasses( ) override{
+
+                std::vector< tardigradeHydra::residualBase* > residuals( 2 );
+
+                elasticity = stressMock( this, 45 );
+
+                plasticity = residualMock( this, 55, 1, stateVariableIndices, plasticParameters );
+
+                residuals[ 0 ] = &elasticity;
+
+                residuals[ 1 ] = &plasticity;
+
+                setResidualClasses( residuals );
+
+            }
+
+    };
+
+    hydraBaseMicromorphicMock hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                                     microDeformation, previousMicroDeformation, gradientMicroDeformation, previousGradientMicroDeformation,
+                                     { }, { },
+                                     previousStateVariables, parameters,
+                                     numConfigurations, numNonLinearSolveStateVariables,
+                                     dimension, configuration_unknown_count,
+                                     tolr, tola, maxIterations, maxLSIterations, lsAlpha );
+
+    residualMock R( &hydra, 55, 1, stateVariableIndices, parameters );
+
+    BOOST_TEST( 1 == *R.getMaxMacroPlasticDeltaNorm( ) );
+
+    BOOST_TEST( 1 == *R.getMaxMicroPlasticDeltaNorm( ) );
+
+    BOOST_TEST( 1 == *R.getMaxMicroGradientPlasticDeltaNorm( ) );
+
+    R.setMaxMacroPlasticDeltaNorm( 1.1 );
+
+    R.setMaxMicroPlasticDeltaNorm( 2.2 );
+
+    R.setMaxMicroGradientPlasticDeltaNorm( 3.3 );
+
+    BOOST_TEST( 1.1 == *R.getMaxMacroPlasticDeltaNorm( ) );
+
+    BOOST_TEST( 2.2 == *R.getMaxMicroPlasticDeltaNorm( ) );
+
+    BOOST_TEST( 3.3 == *R.getMaxMicroGradientPlasticDeltaNorm( ) );
+
+}
+
 BOOST_AUTO_TEST_CASE( test_weakMac, * boost::unit_test::tolerance( DEFAULT_TEST_TOLERANCE ) ){
     /*!
      * Test of the weakened Macaulay bracket
@@ -14930,6 +15128,8 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians2, * boost::unit_test::toler
 
                 plasticity = residualMock( this, 55, 1, stateVariableIndices, plasticParameters );
 
+                plasticity.setUseProjection( false );
+
                 residuals[ 0 ] = &elasticity;
 
                 residuals[ 1 ] = &plasticity;
@@ -14951,6 +15151,8 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians2, * boost::unit_test::toler
     tardigradeHydra::unit_test::hydraBaseTester::updateUnknownVector( hydra, unknownVector );
 
     residualMock R(  &hydra, 55, 1, stateVariableIndices, parameters );
+
+    R.setUseProjection( false );
 
     floatMatrix jacobian( 10, floatVector( unknownVector.size( ), 0 ) );
 
@@ -14985,6 +15187,10 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians2, * boost::unit_test::toler
         residualMock Rp(  &hydrap, 55, 1, stateVariableIndices, parameters );
 
         residualMock Rm(  &hydram, 55, 1, stateVariableIndices, parameters );
+
+        Rp.setUseProjection( false );
+
+        Rm.setUseProjection( false );
 
         floatVector vp = *Rp.get_stateVariableResiduals( );
 
@@ -16411,6 +16617,8 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians_weak2, * boost::unit_test::
 
                 plasticity = residualMock( this, 55, 1, stateVariableIndices, plasticParameters, 0.5, true, a );
 
+                plasticity.setUseProjection( false );
+
                 residuals[ 0 ] = &elasticity;
 
                 residuals[ 1 ] = &plasticity;
@@ -16434,6 +16642,8 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians_weak2, * boost::unit_test::
     floatType a = 0.75;
 
     residualMock R(  &hydra, 55, 1, stateVariableIndices, parameters, 0.5, true, a );
+
+    R.setUseProjection( false );
 
     floatMatrix jacobian( 10, floatVector( unknownVector.size( ), 0 ) );
 
@@ -16468,6 +16678,10 @@ BOOST_AUTO_TEST_CASE( test_setStateVariableJacobians_weak2, * boost::unit_test::
         residualMock Rp(  &hydrap, 55, 1, stateVariableIndices, parameters, 0.5, true, a );
 
         residualMock Rm(  &hydram, 55, 1, stateVariableIndices, parameters, 0.5, true, a );
+
+        Rp.setUseProjection( false );
+
+        Rm.setUseProjection( false );
 
         floatVector vp = *Rp.get_stateVariableResiduals( );
 
@@ -18203,5 +18417,252 @@ BOOST_AUTO_TEST_CASE( test_setdRdT, * boost::unit_test::tolerance( DEFAULT_TEST_
     }
 
     BOOST_TEST( dRdT == *R.getdRdT( ), CHECK_PER_ELEMENT );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_projectSuggestedX, * boost::unit_test::tolerance( DEFAULT_TEST_TOLERANCE ) ){
+
+    // Form a hydra class and residual class
+    floatType time = 1.23;
+
+    floatType deltaTime = 2.34;
+
+    floatType temperature = 3.45;
+
+    floatType previousTemperature = 4.56;
+
+    variableVector deformationGradient = {  1.00157757,  0.00159138,  0.00672005,
+                                            0.01747159,  1.01122277,  0.00555118,
+                                            0.01112217, -0.00885205,  0.99308943 };
+
+    variableVector microDeformation = { 0.99460588, -0.0078411 ,  0.01145249,
+                                       -0.00307139,  0.97798389, -0.00509779,
+                                        0.01189977, -0.01587541,  0.98377259 };
+
+    variableVector gradientMicroDeformation = { -1.35868385e-02, -1.03142977e-02,  6.54880619e-03, -2.03947530e-02,
+                                                -3.31494137e-03, -3.45686183e-03, -3.15745117e-04, -3.70848549e-03,
+                                                -9.38693885e-03, -3.68243465e-03,  1.96694582e-02,  2.22080009e-02,
+                                                 9.18337942e-05,  6.19764759e-03, -1.92190802e-02, -9.13572591e-03,
+                                                -4.25868940e-03,  1.83154579e-02, -1.24772317e-02, -8.48286787e-04,
+                                                 2.42779893e-02,  9.74255963e-04,  5.64472629e-03, -1.89685667e-02,
+                                                 1.63170400e-02,  5.15300642e-03,  2.25340032e-03 };
+
+    floatVector previousDeformationGradient = { 9.94656270e-01,  4.82152400e-02,  3.31984800e-02,  2.81918700e-02,
+         1.02086536e+00, -1.77592100e-02, -2.24798000e-03, -1.28410000e-04,
+         9.77165250e-01 };
+
+    floatVector previousMicroDeformation = { 0.96917405, -0.01777599,  0.00870406, -0.02163002,  0.9998683 ,
+        -0.01669352,  0.03355217,  0.04427456,  1.01778466 };
+
+    floatVector previousGradientMicroDeformation = { 0.05043761,  0.02160516, -0.0565408 ,  0.01218304, -0.05851034,
+         0.00485749, -0.00962607, -0.03455912,  0.04490067,  0.01552915,
+        -0.02878364,  0.00595866,  0.04750406, -0.02377005, -0.05041534,
+        -0.02922214,  0.06280788,  0.02850865, -0.00226005,  0.0146049 ,
+         0.01560184,  0.03224767,  0.05822091, -0.05294424, -0.03518206,
+         0.01831308,  0.03774438 };
+
+    floatVector previousStateVariables = {-0.02495446, -0.00169657,  0.04855598,  0.00194851,  0.01128945,
+       -0.03793713,  0.03263408,  0.01030601,  0.0045068 , -0.01572362,
+       -0.01958792, -0.00829778,  0.01813008,  0.03754568,  0.00104223,
+        0.01693138,  0.00859366,  0.01249035,  0.01746891,  0.03423424,
+       -0.0416805 ,  0.02636828, -0.02563336, -0.0305777 ,  0.0072457 ,
+       -0.04042875,  0.03853268,  0.0127249 ,  0.02234164, -0.04838708,
+        0.00944319,  0.00567852, -0.03410404, -0.03469295,  0.01955295,
+       -0.01812336,  0.01919703,  0.00543832, -0.01110494,  0.04251325,
+        0.034167  , -0.01426024, -0.04564085, -0.01952319, -0.01018143,
+        0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0 };
+
+    floatVector parameters = { 2, 0.53895133, 0.37172145,
+                               2, 0.37773052, 0.92739145,
+                               2, 0.53186824, 0.75454313,
+                               3, 0.95338442, 0.74042148, 0.09916127,
+                               3, 0.38093104, 0.49241325, 0.46187452,
+                               5, 0.82121039, 0.90566759, 0.50466975, 0.04830311, 0.85951495,
+                               3, 0.01166325, 0.05331896, 0.28081774,
+                               3, 0.32982199, 0.60161431, 0.33157768,
+                               5, 0.58881096, 0.11473813, 0.58001078, 0.83382529, 0.3260217 };
+
+    floatVector unknownVector = {  4.31202550e-01, -1.78960429e-01, -6.17986089e-01,  9.34988614e-01,
+                                   3.01500733e-01,  7.30919703e-01, -9.49515284e-01, -4.66188370e-01,
+                                   4.14220065e-03, -8.65102730e-01,  9.86066522e-01, -5.27075208e-01,
+                                  -2.51415635e-01, -5.71976170e-01, -7.89108268e-01, -5.35040429e-01,
+                                  -3.98779729e-01,  2.68884536e-01, -4.37530437e-01, -2.75446478e-01,
+                                  -9.88114313e-01, -2.68561748e-01,  6.77719634e-02, -6.75968326e-01,
+                                   1.94866217e-01, -4.13695063e-01,  2.64100990e-01, -9.47606789e-01,
+                                   7.75186921e-01, -9.67762739e-01, -7.46083938e-01,  5.54324923e-01,
+                                  -9.08209536e-01,  4.21997387e-01,  9.42092281e-01,  7.43365866e-01,
+                                   4.20323303e-01,  9.17019486e-01, -1.40373324e-01,  7.45757829e-01,
+                                  -2.88084664e-01,  8.59527306e-01, -7.02444688e-01,  8.80058030e-01,
+                                   6.65432395e-01,  1.01730274e+00, -1.88038495e-02,  4.82434492e-03,
+                                  -2.41803760e-02,  1.01105922e+00, -2.46131243e-02, -2.07588861e-02,
+                                  -1.37250795e-02,  1.01875623e+00,  9.93178816e-01,  1.99799676e-03,
+                                   3.40516069e-03, -1.37268320e-02,  1.00360734e+00,  8.04758975e-03,
+                                  -1.00877303e-02, -4.06865705e-03,  9.97654446e-01,  2.16175331e-02,
+                                   4.37468737e-03,  2.24126186e-02,  2.80173769e-03,  2.80710425e-05,
+                                  -2.48233895e-02, -9.55547806e-04,  2.13727499e-02, -1.50817155e-02,
+                                  -2.23954433e-02, -4.66105533e-03, -6.38017597e-03,  1.78576529e-02,
+                                  -2.36694442e-02,  2.10074615e-02,  9.04514995e-03,  2.02112997e-02,
+                                   5.37645354e-03,  1.55976656e-02, -8.22280632e-03, -7.52168860e-03,
+                                  -5.50628848e-03,  1.27398541e-02, -6.53544128e-03, -1.28890097e-02,
+                                   2.18834178e-02,  2.04005542e-02,
+                                   0.01, 0.02, 0.03, 0.04, 0.05,
+                                   0.06, 0.07, 0.08, 0.09, 0.10 };
+
+    unsigned int numConfigurations = 2;
+
+    unsigned int numNonLinearSolveStateVariables = 10;
+
+    std::vector< unsigned int > stateVariableIndices = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+
+    unsigned int dimension = 3;
+
+    unsigned int configuration_unknown_count = 45;
+
+    floatType tolr = 1e-2;
+
+    floatType tola = 1e-3;
+
+    unsigned int maxIterations = 24;
+
+    unsigned int maxLSIterations = 45;
+
+    floatType lsAlpha = 2.3;
+
+    class stressMock : public tardigradeHydra::residualBaseMicromorphic{
+
+        public:
+
+            using tardigradeHydra::residualBaseMicromorphic::residualBaseMicromorphic;
+
+    };
+
+    class residualMock : public tardigradeHydra::micromorphicDruckerPragerPlasticity::residual{
+
+        public:
+
+            using tardigradeHydra::micromorphicDruckerPragerPlasticity::residual::residual;
+
+    };
+
+    class hydraBaseMicromorphicMock : public tardigradeHydra::hydraBaseMicromorphic{
+
+        public:
+
+            using tardigradeHydra::hydraBaseMicromorphic::hydraBaseMicromorphic;
+
+            stressMock elasticity;
+
+            residualMock plasticity;
+
+            std::vector< unsigned int > stateVariableIndices = { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 };
+
+            floatVector plasticParameters = { 2, 0.53895133, 0.37172145,
+                                              2, 0.37773052, 0.92739145,
+                                              2, 0.53186824, 0.75454313,
+                                              3, 0.95338442, 0.74042148, 0.09916127,
+                                              3, 0.38093104, 0.49241325, 0.46187452,
+                                              5, 0.82121039, 0.90566759, 0.50466975, 0.04830311, 0.85951495,
+                                              3, 0.01166325, 0.05331896, 0.28081774,
+                                              3, 0.32982199, 0.60161431, 0.33157768,
+                                              5, 0.58881096, 0.11473813, 0.58001078, 0.83382529, 0.3260217 };
+
+        private:
+
+            using tardigradeHydra::hydraBaseMicromorphic::setResidualClasses;
+
+            virtual void setResidualClasses( ) override{
+
+                std::vector< tardigradeHydra::residualBase* > residuals( 2 );
+
+                elasticity = stressMock( this, 45 );
+
+                plasticity = residualMock( this, 55, 1, stateVariableIndices, plasticParameters );
+
+                residuals[ 0 ] = &elasticity;
+
+                residuals[ 1 ] = &plasticity;
+
+                setResidualClasses( residuals );
+
+            }
+
+    };
+
+    hydraBaseMicromorphicMock hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                                     microDeformation, previousMicroDeformation, gradientMicroDeformation, previousGradientMicroDeformation,
+                                     { }, { },
+                                     previousStateVariables, parameters,
+                                     numConfigurations, numNonLinearSolveStateVariables,
+                                     dimension, configuration_unknown_count,
+                                     tolr, tola, maxIterations, maxLSIterations, lsAlpha );
+
+    residualMock R( &hydra, 55, 1, stateVariableIndices, parameters );
+
+    floatVector trialX = { -0.40474133, -1.31581257, -0.24799601, -0.66822701,  1.70799134,
+                            0.13089057, -0.17216613, -0.00864461,  0.80232873,  0.13049679,
+                            0.3705609 ,  1.04375356,  2.1860939 ,  0.65654621, -0.91634756,
+                           -1.54474396, -0.17666782,  1.65885133,  2.29853222,  0.53338233,
+                            0.69052704,  0.78619126,  1.18703091,  1.83891347, -1.20737302,
+                            0.48385701, -0.75505922, -1.68384194, -0.01766042,  1.91917888,
+                            1.34376976, -0.53305317,  0.46286124,  1.48304898,  1.36610329,
+                            0.1893202 , -1.42179529, -2.49648211, -0.08692722, -1.2776921 ,
+                            2.28801669, -2.00114005,  1.51340049, -2.17481968, -0.96377428,
+                           -0.70299002, -1.47433365,  3.44466764, -0.6873039 , -1.30451035,
+                           -2.0131308 , -1.62654174,  0.68615533, -0.41187571,  4.54269611,
+                           -0.28936713,  1.9595505 ,  3.64415161, -0.21041396, -3.00618013,
+                           -3.57395263,  4.55948367,  4.62129283,  0.73171675, -2.97715061,
+                            2.36464984,  0.04360128, -3.57926391,  1.66380595, -2.76950485,
+                           -2.16234317, -3.68731285, -3.51451628,  2.71405279,  4.41692098,
+                            4.03053223,  0.12819252,  2.11006335, -0.13984174, -2.25057865,
+                           -3.81652212, -3.42465008, -4.11456265,  4.52170914, -3.62536239,
+                           -4.50230405,  3.28010857, -0.13044935, -1.88583842,  0.88096962,
+                           -0.11414338,  0.07622781, -0.11073456, -0.48607211, -0.02456945,
+                            0.46456126, -0.27655936, -0.383391  ,  0.26888644,  0.35932029 };
+
+    floatVector Xp = { 2.37179207,  2.04571891,  1.43340891,  1.56210501,  0.29650219,
+                      -1.81653588, -1.40387454, -1.2059766 ,  1.98287363, -1.17985447,
+                      -0.31947747,  1.67487249, -2.2389102 , -1.02632147, -0.04017631,
+                       1.53355464,  1.77147411, -0.85353032, -0.54185951, -1.91388754,
+                       1.35593112,  2.27108171,  1.20009559, -2.32963703, -1.36833298,
+                       1.57008187, -0.68035577,  1.62855286,  2.08180914, -0.22511169,
+                       2.23598071,  1.13591538,  1.8935963 ,  1.02065205, -1.52995918,
+                       1.86641792, -2.26013904, -1.3023638 , -1.95985108, -2.43277805,
+                       1.30597982,  1.15786049, -1.44268755, -2.4201092 , -0.35044801,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.        ,
+                       0.        ,  0.        ,  0.        ,  0.        ,  0.         };
+
+    floatVector answer = { -0.40474133, -1.31581257, -0.24799601, -0.66822701,  1.70799134,
+                            0.13089057, -0.17216613, -0.00864461,  0.80232873,  0.13049679,
+                            0.3705609 ,  1.04375356,  2.1860939 ,  0.65654621, -0.91634756,
+                           -1.54474396, -0.17666782,  1.65885133,  2.29853222,  0.53338233,
+                            0.69052704,  0.78619126,  1.18703091,  1.83891347, -1.20737302,
+                            0.48385701, -0.75505922, -1.68384194, -0.01766042,  1.91917888,
+                            1.34376976, -0.53305317,  0.46286124,  1.48304898,  1.36610329,
+                            0.1893202 , -1.42179529, -2.49648211, -0.08692722, -1.2776921 ,
+                            2.28801669, -2.00114005,  1.51340049, -2.17481968, -0.96377428,
+                           -0.1433584 , -0.30065593,  0.70245954, -0.14015958, -0.26602443,
+                           -0.41053102, -0.33169521,  0.13992536, -0.08399244,  0.45014875,
+                           -0.02867422,  0.19417746,  0.3611094 , -0.02085052, -0.29789099,
+                           -0.35415319,  0.45181228,  0.45793712,  0.04786126, -0.19473407,
+                            0.15467067,  0.00285194, -0.23411803,  0.10882879, -0.18115205,
+                           -0.14143788, -0.24118546, -0.22988291,  0.17752496,  0.2889088 ,
+                            0.26363528,  0.00838501,  0.13801829, -0.00914698, -0.14720933,
+                           -0.24963698, -0.22400481, -0.26913167,  0.29576294, -0.23713331,
+                           -0.29449366,  0.21455041, -0.00853263, -0.12335183,  0.05762382,
+                            0.        ,  0.07622781,  0.        ,  0.        ,  0.        ,
+                            0.46456126,  0.        ,  0.        ,  0.26888644,  0.35932029 };
+
+    R.projectSuggestedX( trialX, Xp );
+
+    BOOST_TEST( trialX == answer, CHECK_PER_ELEMENT );
 
 }